Added case-insensitive function #24
Created 3 more functions that checks if the number of rows matches the
different combinations in the database. Added conditionals to
exports.saveRequest to see which error message will appear when the
user enters in the incorrect combination of city, state, and zip code.
The error messages is not complete, need more tests.

Fixed case-insensitive issue in db-call #24

Created a function that capitalizes every word in the string and
applied it to the city parameter on every city-state-zip functions
that makes a db call. Prevoiusly, we would get an error if the user
filled in city field as "chicago" instead of "Chicago" (capitalized)
because the lower cased "chicago" does not exist in the database.