Some translations removed & rearranged #24
i18n auto-generated files were not committed when changes were
made from previous commits. Duplicate messages are removed and some
messages are rearranged because of the merging of the two branches.
Previously, when the i18n files reloads and a new request is made
at the same time a blank screen sometimes appears instead of the
city-state-zip error message.